+++ R/summary.tableby.R
@@ -130,7 +130,7 @@ as_data_frame_summary_tableby <- function(df, totals, hasStrata, term.name, cont
   df$test <- NULL
   df$variable.type <- NULL
   if(!control$test) df$p.value <- NULL
-  if(!control$total && !identical(control$stats.labels$overall, "Total")) df$Total <- NULL
+  if(!control$total && !identical(control$stats.labels$overall, "Total")) df[[control$stats.labels$total]] <- NULL
 
   #### Format if necessary ####
   if(!is.null(width))